docs: fix typos and standardize build tags
Typos fixed:
- doc.go: "mutliple" -> "multiple"
- apply.go: "no public key" -> "No signature" (capitalization)
- verifier.go: "ecsda" -> "ecdsa"
- encoding.go: remove redundant uint64(uint64()) conversion

Build tags standardized to //go:build syntax:
- osext_procfs.go: linux || netbsd || openbsd || solaris || dragonfly
- osext_sysctl.go: darwin || freebsd

These changes improve code quality and consistency.

+++ internal/binarydist/encoding.go
@@ -31,7 +31,7 @@ func (signMagLittleEndian) Uint64(b []byte) uint64 {
 	}
 	// Return bitwise representation of int64 as uint64
 	// #nosec G115 - Required for bsdiff sign-magnitude encoding
-	return uint64(uint64(y))
+	return uint64(y)
 }
 
 func (signMagLittleEndian) PutUint64(b []byte, v uint64) {
